When checking map key, fail if not a Python string
Previously, there was a logic error where a key for an Avro map
was considered invalid if it passed the PyUnicode_CheckExact call.
A valid key should always successfully pass that check, so negate
that check to detect failure.

In addition, only Python3 actually has unicode keys, and Python2 uses
str values.  Use the private _PyUnicode_CheckExact macro in order to
have Python2/Python3 compatibilty.

Add a basic unit test for testing the encoding of the map type.

fixes #3