Add support for optional Fixed/Bytes in Python3
The python_to_union function needs another manual type check
for PyBytes in order to support Avro Bytes and Fixed types under
Python 3.  This line is not reached under Python 2 in the case of
Bytes and Fixed types due to the check right above.  Add in a new
unit test for this too.

fixes #13