Fixes #23
The avro float is a 32-bit float so I had to adjust how it currently returns that
information to Python which only has a 64-bit float.  The problem causing the segfault was
only related to improper NULL handling in the `lookup` function (there was no entry for
'float' so it got to the last item and didn't check the NULL), however, to have proper
float and double support I needed to update the conversion functions and add something to
round the 32-bit avro float to reduce the noise introduced in the conversion (see
[here](https://stackoverflow.com/questions/3498192/c-convert-double-to-float-preserving-decimal-point-precision)).
I also included a test to look for unhandled union lookups to effectively test for what was
causing the segfault.