Iterate over fields in record to validate
When performing validation of a record against a candidate schema,
one should check the fields in the record itself against the
schema.  If one checks the fields in the schema against the record
instead, then there may be fields present in the record which get
ignored and thus dropped.  This is particularly important when
there is a union of record types where one record type has a subset
of fields of the other record type.

Addresses #27