Add schema resolution support
Using AvroC's resolving writer/reader allows the implementation
of schema resolution. In the case where a new "reader schema"
is used while reading avro records, this schema will be applied
rather than the originating schema (confusingly referred to as
the "writer schema"). In cases when writing avro records the
so-called "reader schema" is used to resolve to the new schema,
allowing such things like promotion of types (where valid).
This ends up fixing #30 where the need is to skip fields that
are being resolved for performance reasons.